Support non-breaking content (`<nobr>` and `white-space: nowrap`) (#240)
This patch supports non-breaking content in the JavaScript `HTMLProcessor`. The non-breaking content provides context for the Parser but the content is not breakable.

Contents with `white-space: nowrap` and `white-spce: pre` are handled as non-breaking content. The `<nobr>` element is also supported; this is useful for disconnected tree when all styles including `white-space` are not computed.